move package to pkg
