+++ pkg/soap/parse_test.go
